memH: Improve MemLink interface and make MemNICBase class its own file
