+++ src/sst/elements/memHierarchy/cacheEventProcessing.cc
@@ -497,30 +497,7 @@ void Cache::init(unsigned int phase) {
 
 void Cache::setup() {
     // Check that our sources and destinations exist or configure if needed
-    
-    std::set<MemLinkBase::EndpointInfo> * names = linkUp_->getSources();
-
-    if (names->empty()) {
-        std::set<MemLinkBase::EndpointInfo> srcNames;
-        if (upperLevelCacheNames_.empty()) upperLevelCacheNames_.push_back(""); // TODO is this a carry over from the old init or is it needed to avoid segfaults still?
-        for (int i = 0; i < upperLevelCacheNames_.size(); i++) {
-            MemLinkBase::EndpointInfo info;
-            info.name = upperLevelCacheNames_[i];
-            info.addr = 0;
-            info.id = 0;
-            info.region.setDefault();
-            srcNames.insert(info);
-        }
-        linkUp_->setSources(srcNames);
-    }
-    names = linkUp_->getSources();
-    if (names->empty()) 
-        out_->fatal(CALL_INFO, -1,"%s did not find any sources\n", getName().c_str());
-
-    names = linkDown_->getDests();
-    if (names->empty())
-        out_->fatal(CALL_INFO, -1, "%s did not find any destinations\n", getName().c_str());
-
+   
     linkUp_->setup();
     if (linkUp_ != linkDown_) linkDown_->setup();
